Move `Username()` from `api.go` to `tsh.go`
